# Don't allocate strings for unchanged ParameterStatus