--- a/src/Npgsql/NpgsqlConnector.cs
+++ b/src/Npgsql/NpgsqlConnector.cs
@@ -148,6 +148,11 @@ namespace Npgsql
         internal readonly Dictionary<string, string> PostgresParameters;
 
         /// <summary>
+        /// Holds all run-time parameters in raw, binary format for efficient handling without allocations.
+        /// </summary>
+        readonly List<(byte[] Name, byte[] Value)> _rawParameters = new List<(byte[] Name, byte[] Value)>();
+
+        /// <summary>
         /// The timeout for reading messages that are part of the user's command
         /// (i.e. which aren't internal prepended commands).
         /// </summary>
@@ -1050,7 +1055,7 @@ namespace Npgsql
                 case BackendMessageCode.CloseComplete:
                     return CloseCompletedMessage.Instance;
                 case BackendMessageCode.ParameterStatus:
-                    HandleParameterStatus(buf.ReadNullTerminatedString(), buf.ReadNullTerminatedString());
+                    ReadParameterStatus(buf.GetNullTerminatedBytes(), buf.GetNullTerminatedBytes());
                     return null;
                 case BackendMessageCode.NoticeResponse:
                     var notice = new PostgresNotice(buf);
@@ -1973,8 +1978,31 @@ namespace Npgsql
 
         #region Misc
 
-        void HandleParameterStatus(string name, string value)
+        void ReadParameterStatus(ReadOnlySpan<byte> incomingName, ReadOnlySpan<byte> incomingValue)
         {
+            (byte[], byte[]) entry = default;
+            foreach (var e in _rawParameters)
+            {
+                if (incomingName.SequenceEqual(new ReadOnlySpan<byte>(e.Name)))
+                {
+                    entry = (e.Name, e.Value);
+                    break;
+                }
+            }
+
+            if (entry == default((byte[], byte[])))
+            {
+                entry = (incomingName.ToArray(), incomingValue.ToArray());
+                _rawParameters.Add(entry);
+            }
+            else if (!incomingValue.SequenceEqual(new ReadOnlySpan<byte>(entry.Item2)))
+                entry.Item2 = incomingValue.ToArray();
+            else
+                return;
+
+            var name = TextEncoding.GetString(entry.Item1);
+            var value = TextEncoding.GetString(entry.Item2);
+
             PostgresParameters[name] = value;
 
             switch (name)