--- a/src/Npgsql/NpgsqlReadBuffer.cs
+++ b/src/Npgsql/NpgsqlReadBuffer.cs
@@ -435,6 +435,20 @@ namespace Npgsql
             return result;
         }
 
+        public ReadOnlySpan<byte> GetNullTerminatedBytes()
+        {
+            int i;
+            for (i = ReadPosition; Buffer[i] != 0; i++)
+            {
+                Debug.Assert(i <= ReadPosition + ReadBytesLeft);
+            }
+            Debug.Assert(i >= ReadPosition);
+
+            var result = new ReadOnlySpan<byte>(Buffer, ReadPosition, i - ReadPosition);
+            ReadPosition = i + 1;
+            return result;
+        }
+
         #endregion
 
         #region Misc