Closes #2236