@roji 2019-03-21
# Before
``` ini

BenchmarkDotNet=v0.11.3, OS=ubuntu 18.10
Intel Xeon W-2133 CPU 3.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.0.100-preview-009812
  [Host]     : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT
  DefaultJob : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT


```
|                  Method | StatementsToSend |           Mean |         Error |        StdDev |         Median |        Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------ |----------------- |---------------:|--------------:|--------------:|---------------:|------------:|------------:|------------:|------------:|--------------------:|
| **OpenCloseSameConnection** |                **1** |   **120,046.3 ns** |  **7,552.643 ns** | **22,269.138 ns** |   **125,735.4 ns** |     **8,330.1** |      **0.2441** |           **-** |           **-** |              **1116 B** |

# After

``` ini

BenchmarkDotNet=v0.11.3, OS=ubuntu 18.10
Intel Xeon W-2133 CPU 3.60GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=3.0.100-preview-009812
  [Host]     : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT
  DefaultJob : .NET Core 3.0.0-preview-27122-01 (CoreCLR 4.6.27121.03, CoreFX 4.7.18.57103), 64bit RyuJIT


```
|                  Method | StatementsToSend |           Mean |         Error |        StdDev |         Median |        Op/s | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------------------------ |----------------- |---------------:|--------------:|--------------:|---------------:|------------:|------------:|------------:|------------:|--------------------:|
| **OpenCloseSameConnection** |                **1** |   **124,295.6 ns** |  **6,288.555 ns** | **18,541.947 ns** |   **127,763.3 ns** |     **8,045.3** |           **-** |           **-** |           **-** |               **921 B** |